scraper/requirements.txt
playwright==1.44.0
apscheduler==3.10.4
pika==1.3.2
pyyaml==6.0.1
watchdog==4.0.0
structlog==24.1.0
